[Fix] View Controller 안의 상수 부분 ConstantValue파일로 이동
